Initial commit for wip v2
This commit introduces the initial changes for the second version of the website. It includes new files for settings, application styles, and main application component. It also includes modifications to the main index file. Several components and pages from the previous version have been removed as part of this update.